# project/application/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Service API"
  description: "Flask-based REST API for managing customers, mechanics, service tickets, and inventory."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

paths:

  /customers:
    get:
      summary: "List all customers"
      parameters:
        - name: page
          in: query
          type: integer
          required: false
          default: 1
        - name: per_page
          in: query
          type: integer
          required: false
          default: 10
      responses:
        200:
          description: "Paginated list of customers"
    post:
      summary: "Create a new customer"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CustomerCreate"
      responses:
        201:
          description: "Customer created"

  /customers/login:
    post:
      summary: "Customer login"
      description: "Authenticate customer and return JWT token"
      parameters:
        - in: body
          name: credentials
          schema:
            $ref: "#/definitions/Login"
      responses:
        200:
          description: "JWT token returned"

  /customers/{customer_id}:
    get:
      summary: "Get a customer by ID"
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Customer object"
        404:
          description: "Not found"
    put:
      summary: "Update customer details"
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          schema:
            $ref: "#/definitions/CustomerUpdate"
      responses:
        200:
          description: "Updated customer"
    delete:
      summary: "Delete a customer"
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Customer deleted"

  /mechanics:
    get:
      summary: "List all mechanics"
      responses:
        200:
          description: "List of mechanics"
    post:
      summary: "Create a mechanic"
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/Mechanic"
      responses:
        201:
          description: "Mechanic created"

  /mechanics/{mechanic_id}:
    get:
      summary: "Get mechanic by ID"
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Mechanic found"
    put:
      summary: "Update mechanic info"
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          schema:
            $ref: "#/definitions/Mechanic"
      responses:
        200:
          description: "Updated mechanic"
    delete:
      summary: "Delete mechanic"
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Deleted mechanic"

  /inventory:
    get:
      summary: "List inventory parts"
      responses:
        200:
          description: "All inventory parts"
    post:
      summary: "Add new inventory part"
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/InventoryItem"
      responses:
        201:
          description: "Inventory item created"

  /inventory/{part_id}:
    get:
      summary: "Get part by ID"
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Part details"
    put:
      summary: "Update part info"
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          schema:
            $ref: "#/definitions/InventoryItem"
      responses:
        200:
          description: "Updated part"
    delete:
      summary: "Delete part"
      parameters:
        - name: part_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Part deleted"

  /service_tickets:
    get:
      summary: "List all service tickets"
      responses:
        200:
          description: "List of service tickets"
    post:
      summary: "Create a service ticket"
      parameters:
        - in: body
          name: body
          schema:
            $ref: "#/definitions/TicketCreate"
      responses:
        201:
          description: "Ticket created"

  /service_tickets/{ticket_id}:
    get:
      summary: "Get a specific ticket"
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Ticket details"
    put:
      summary: "Update a service ticket"
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          schema:
            $ref: "#/definitions/TicketUpdate"
      responses:
        200:
          description: "Updated ticket"
    delete:
      summary: "Delete a service ticket"
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "Ticket deleted"

# --------------------------------
# Models / Schemas
# --------------------------------
definitions:
  CustomerCreate:
    type: object
    required: [name, email, phone, password]
    properties:
      name: { type: string, example: "Ash" }
      email: { type: string, example: "ash@example.com" }
      phone: { type: string, example: "555-111-2222" }
      password: { type: string, example: "Secret123!" }

  CustomerUpdate:
    type: object
    properties:
      name: { type: string }
      email: { type: string }
      phone: { type: string }

  Login:
    type: object
    required: [email, password]
    properties:
      email: { type: string, example: "ash@example.com" }
      password: { type: string, example: "Secret123!" }

  Mechanic:
    type: object
    required: [name, email, phone, salary]
    properties:
      name: { type: string, example: "Brock" }
      email: { type: string, example: "brock@example.com" }
      phone: { type: string, example: "555-444-3333" }
      salary: { type: number, example: 4500 }

  InventoryItem:
    type: object
    required: [name, price]
    properties:
      name: { type: string, example: "Brake Pad" }
      price: { type: number, example: 99.99 }

  TicketCreate:
    type: object
    required: [VIN, service_date, service_desc, customer_id]
    properties:
      VIN: { type: string, example: "1HGCM82633A004352" }
      service_date: { type: string, format: date, example: "2025-09-01" }
      service_desc: { type: string, example: "Brake replacement" }
      customer_id: { type: integer, example: 1 }
      mechanic_ids:
        type: array
        items: { type: integer }

  TicketUpdate:
    type: object
    properties:
      service_desc: { type: string }
      mechanic_ids:
        type: array
        items: { type: integer }